\name{fcov}
\alias{fcov}
\title{fcov}
\description{
 Calculeaza covarianta dintre v.a. X si Y.}
\usage{fcov(X,Y,repCom)}

\arguments{
  \item{X}{Functia de masa a v.a X.}
  \item{Y}{Functia de masa a v.a Y.}
  \item{repCom}{Repartitia comuna a v.a X si Y.}
}
\details{}
\value{covarianta dintre v.a. X si Y.}
